Added trap for EXIT signal
Bash does not trigger the ERR-signal for "nounset" errors, but just
quits to script by the special EXIT-signal.

Also, provide pkgbox_die() with some more info from trap handler.
pkgbox_die() now checks for a zero-exitcode and displays the message and
frames as "warning" and exits also with success. All traps are removed
before the final call to exit.

The main script (pkgbox) at the very end also removes the EXIT-trap and
exits with success.